Treat `form > legend` contents as Markdown
But remove the unnecessary `<p>` wrapper with this inconvenient hack
from Stack Overflow:

https://stackoverflow.com/questions/30824395/jekyll-how-can-i-make-avoid-a-paragraph-to-be-added-on-a-yaml-frontmatter-mark

I looked around A LOT for alternatives and concluded that everyone is
either using this hack or using a custom plugin. I'm not hand-rolling a
Jekyll plugin for this... and going through the steps to publish without
the default GitHub Pages build process (aka GitHub Actions).